requirements.txt
twitchio
requests
python-dotenv
beautifulsoup4
cloudscraper
discord.py